+++ librecipebook/src/data/SortOrder.cpp
@@ -19,20 +19,20 @@ const recipebook::Category& recipebook::SortOrder::getItemAt(quint32 i) const
 		throw QException();
 	}
 
-	return *m_Categories.at(i);
+    return *m_Categories.at((int)i);
 }
 
 void recipebook::SortOrder::moveCategory(const Category& category, quint32 newPos)
 {
-	quint32 oldPos = m_Categories.indexOf(&category);
-	m_Categories.move(oldPos, newPos);
+    int oldPos = m_Categories.indexOf(&category);
+    m_Categories.move(oldPos, (int)newPos);
 }
 
 quint32 recipebook::SortOrder::getIndex(const Category& category) const
 {
 	for(quint32 i = 0; i < (quint32)m_Categories.size(); ++i)
 	{
-		if(m_Categories.at(i) == &category)
+        if(m_Categories.at((int)i) == &category)
 		{
 			return i;
 		}
